raphael a presque finni juste le makefile et je push

+++ srcs/requirements/mariadb/lala.sh
@@ -6,11 +6,19 @@ export DB_USER=rkhelif
 export DB_PASS=rkhelif
 
 set -xv
+RAPH=/lala.sql
+if [ -f "$RAPH" ]
+then
 #creation de la base de donner
 service mysql start
 #mysql -V
 envsubst < /lala.sql > /lala2.sql
-mysql -Dmysql < /lala2.sql | true
+mysql < /lala2.sql
+rm -rf /lala.sql
+rm -rf /lala2.sql
+service mysql stop
+fi
+exec mysqld_safe
 
 #echo "CREATE DATABASE IF NOT EXISTS $DB_NAME;" | mysql -u root -p$DB_ROOT_PASS
 #creation d'un nouvel utilisateur qui peut utiliser la base de donner
@@ -19,7 +27,3 @@ mysql -Dmysql < /lala2.sql | true
 #echo "GRANT ALL PRIVILEGES on *.* to 'root'@'localhost' identified by '$DB_ROOT_PASS';" | mysql -u root -p$DB_ROOT_PASS
 #echo "ALTER USER 'root'@'localhost' IDENTIFIED BY '$DB_ROOT_PASS';" | mysql -u root -p$DB_ROOT_PASS
 #echo "FLUSH PRIVILEGES;" | mysql -u root -p$DB_ROOT_PASS
-# sleep 4;
-service mysql stop | echo -n ""
-
-exec mysqld_safe
